Move all initialization related types into a dedicated module.
